выполнил корректировки 1
