+++ src/components/burger/ingredients/Categories/categories.jsx
@@ -1,7 +1,7 @@
 import styles from './categories.module.css';
 import { } from '@ya.praktikum/react-developer-burger-ui-components'
 import classNames from 'classnames';
-import Tabmenu from '../Tab-menu/tab-menu';
+import Tabmenu from '../tab-menu/tab-menu';
 import PropTypes from 'prop-types';
 
 export default function Categories(props) {
